+++ packages/orq-rc/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.2.0-rc.14",
-  genVersion: "2.524.1",
-  userAgent: "speakeasy-sdk/typescript 3.2.0-rc.14 2.524.1 2.0 @orq-ai/node",
+  sdkVersion: "3.2.0-rc.17",
+  genVersion: "2.529.2",
+  userAgent: "speakeasy-sdk/typescript 3.2.0-rc.17 2.529.2 2.0 @orq-ai/node",
 } as const;