+++ packages/orq-rc/src/models/operations/fileget.ts
@@ -145,7 +145,7 @@ export const FileGetResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-02-25T12:28:13.397Z",
+    "2025-02-26T10:55:48.615Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -179,7 +179,7 @@ export const FileGetResponseBody$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-02-25T12:28:13.397Z"))
+  created: z.date().default(() => new Date("2025-02-26T10:55:48.615Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {